[Refactor] Moving and rename some code
