+++ twitter_connector/alembic.ini
@@ -29,7 +29,7 @@ script_location = migrations
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = postgresql://dra@:5432/pypular
+sqlalchemy.url = postgresql://@:5432/pypular
 
 
 # Logging configuration
