+++ twitter_connector/listeners.py
@@ -1,7 +1,7 @@
 import tweepy
-import requests
 
-import api.database as db
+import twitter_connector.models as db
+from twitter_connector.utils import get_expanded_url
 
 import json
 import logging
@@ -162,17 +162,3 @@ class DBListener(tweepy.StreamListener):
 
     def on_error(self, status):
         logger.error(status)
-
-
-def get_expanded_url(url, max_retries=3):
-    # todo: add error handling
-    for retry in range(max_retries):
-        try:
-            req = requests.head(url, allow_redirects=True, timeout=10)
-        except:
-            logger.error('Unable to connect to url: %s', url)
-            logger.info('Retrying the connection. Attempt %s out of %s' % (
-                retry + 1, max_retries))
-        else:
-            return req.url or 'invalid_url'  # TODO: need some verification
-    return url
